Mark "void Boot()" as exported
This marks two functions with HXCPP_EXTERN_CLASS_ATTRIBUTES to make it
easier to use hxcpp as a DLL in Windows. Currently it almost is as
simple as just building all of hxcpp with HXCPP_DLL_EXPORT and defining
HXCPP_DLL_IMPORT when using it, except that there are linker errors
involving "void Boot()" and "int __hxcpp_irand(int inMax)" since they
lack the necessary declspec attributes.