Improved error reporting
Internally hxcpp does a search first on MSVC then followed by Mingw. 

In cases when we dont have either MSVC or Mingw compilers in Windows environment, then tried to do

    hxcpp Build.xml

We get an error related to Mingw. Instead of an error that compilers could not be setup.

This pull will cause error reporting to default to

    Could not setup any C++ compiler, please install or reinstall a valid C++ compiler

Then if -Dmingw is indicated explicitly

    Could not guess MINGW_ROOT (tried $guesses) - please set explicitly