WinRT fixes: threads, 64bit, static compile, optional MD abi
Removed some code that is available in Windows 10 UWP but was not
previusly available in Windows Store 8 using (_WIN32_WINNT ==
_WIN32_WINNT_WIN8)

Encapsulated code for store-api with   defined(__cplusplus_winrt) so it
can compile with the -MD abi (-DABI="-MD"). so you can compile your
static libs more similar to C++ and only use a store-api main in a
Main.cpp using ZW abi.