WinRT. Toolchain and setup adjustments
I had to add a HXCPP_WINRT_MD flag since there seems to be no way to
check the value of "ABI" (like, if ABI is "-MD", "-MT" or "-ZW"). This
leaves the defult of winrt static compilation to ZW, but allows optional
MD compilation more easily. Not sure if also works for non-static_link.

Also I added HXCPP_HACK_PDBSRV to setup, to fix a compile error