Android ndk fix
Fix error when requested "NDKVXX" not found. Fix conversion of slashes
in directories.
It considers ANDROID_SDK/ndk-bundle to be as if it where inside
ANDROID_NDK_DIR and considers using the newer version in the same way.
If ANDROID_NDK_DIR is not set, it will look only for
ANDROID_SDK/ndk-bundle