Handle .length and .nfields on empty result sets
Forward the number of changed rows for statements other than SELECT.
This prevents an exception to be thrown with SQLite,[1] and also matches
the corresponding Neko documentation[2].

Both hx_sqlite_result_get_length and _hx_mysql_result_get_nfields are
changed to work in these scenarios.  The latter change fixes a different
yet related error with MySQL, where .nfields would raise a similar
exception.

Finally, this also fixes an issue with running the MySQL tests locally,
as the 'hxcpp' test database will be reused and has to be cleared.

[1] Sqlite cnx.request return error! (hxcpp#859)
[2] Neko SQLite API (https://nekovm.org/doc/view/sqlite/)