Add warning if destination flag is ignored
If the current target does not produce a file, the -Ddestination flag is
currently ignored without warning.

This situation may occur commonly when 'default' points to another
target, but doesn't produce a file itself. A warning will now be
displayed:

Warning: Target 'default' does not output a file, so 'destination' has
been ignored