[cppia] Respect jit exceptions in interp mode
When running in interp mode, native c++ exception handling is used, but
when running jit mode, jumps are used along with ctx->exception.

This means that an exception thrown from jit mode is not respected in
interp mode. This can cause problems with local functions (which are
interpreted even in jit mode). When a local function calls a compiled
function that throws an exception, the exception is ignored until we
return back into a compiled function.

This patch fixes the problem by checking ctx->exception when running in
interp mode, to make sure that no exception has been thrown from a jit
compiled function.