Prevent closing socket handle twice
If the socket handle has been closed once, it no longer belongs to us
and it is unsafe to close it again as it could have been assigned
somewhere else in the program. On android for example, this is detected
at runtime by fdsan and causes a hard crash.

We could also consider throwing an error on the second close, however,
for now this is already a step forward.