Fix haxelib path fallback functionality on windows
The check on line 134 would never work properly on windows, because the
line had a CR character so FileSystem.exists returned false. This meant
that if a library was resolved via a name that doesn't exactly match the
`-D` flag set by haxelib, e.g. Hxcpp instead of hxcpp or via an alias,
then it would be reported as missing.