Avoid duplicate string iteration in TCopyString
TConvertToUTF8 treats strings as null terminating if the length is 0, so
rather than calculating the size manually we can set it to 0 if it the
input to TCopyString is -1.

For GCStringDup, -1 is handled as null terminating, but 0 is treated as
a 0 length string. We can pass the length argument straight in without
modification there.

This avoids duplicate loops in both the TConvertToUTF8 and GCStringDup
cases.