Fix String::create(_, 0) without HX_SMART_STRINGS
When smart strings are enabled, this always returns an empty string,
however, if they are disabled this currently has different behaviour
depending on the type of character passed in.

This fixes the behaviour so that it also returns an empty string for
wchar_t and char16_t strings when smart strings are disabled.