[cppia] Fix haxe overridden scriptable methods (#1273)
* [cppia] Add flag for scriptable overridden methods

This allows overridden methods to be marked so that they do not add an
addtional slot to the vtable, which has previously caused bugs due to
incorrect vtable indices.

* [cppia] Infer overrides from matching method name

For haxe versions that do not tag overridden methods in
__scriptableFunctions tables, we can infer overrides at runtime by
comparing method names.

* [cppia] Swap scriptable func argument order

This way is a bit more backwards compatible. An overridden method will
always have a super version provided.